Add ZGO routine to output ZWR files organized by FileMan
Walk the global node structure.  Track FileMan "File Global Roots" to
output global nodes in each FileMan file in corresponding ZWR files.
Name the output files as $G+$F.zwr where $G is the global name and $F is
the FileMan file name (translated to a valid host file name).

Currently this works only in Cache because the API to get a list of all
globals and to deal with host files is specific to the M implementation.