Update the testing structure & use a clean Cache
Change the testing structure to glob the VistA FOIA release for the package
names and the routines contained in them to create a test for each Package.
Change the AutoHotkey and Expect scripts to account for the new variables.
Provide the user an option to use the build step to import routines and
globals into the Cache.dat.  Use the GTMEnvironment.sh to source the
environment within the Nightly.sh, rather than having two sets of the same
commands

Change-Id: I58ab2b36eeb244dd2c13715567d4ff1ef1a02b07