EHN:change to use the Cache telnet script.
Added a CacheXINDEXScript.in as input template for
Cache telnet scripts. Changed WinRoutine.cmake.in to use
Cterm instead of CControl to talk to Cache.
Changed CMakeLists.txt to remove the checking for AutoHotKey
and replace CControl with CTerm.

Requirement: In order to be able to run the script

1. you need to launch Cache Managements Portal and go to
System->Security Management->Service to switch on %Service_telnet
by checking enabled checkbox and save. Also verify that telnet port
is set to 23 via
Configuration->Device Settings->Telnet Settings

2. Also please change the $IO variable for TELNET to |TNT|
via Cache Terminal

Change-Id: Ib48c83cd9d814b87d9d8c985eaecd242ba24449c