Adds capability to import OSEHRA code nightly for GT.M
Change the ImportRG to work on a GT.M install.  Removes the database.dat
from the vista folder and removes all routines from the R directory.  Pack
the routines and globals like before, and then create the new database.dat
and import the routines and globals using new expect scripts.  Change the
CMakeLists.txt to have a clean database option and configure the expect
scripts and the cmake file.  The cmake files uses cmake commands to remove
the old files.  Also, include error checking macro to display a fatal error message
if any of the steps should fail to start

Change-Id: I734fd9155fdff3abef09ebba48ab9a307b159f69