Improve Error reporting of results
Change the configured CMake file to include an Error_Variable output on each
execute process.  This will be used to ensure that the execute process
command is executed.  If it returns that it didn't run, send a fatal error
saying that the process failed.  Also put test_passed variable used in
CommonFunctions.cmake to tell when the test fails/passes.  Because of this
addition, remove the FAIL_REGULAR_EXPRESSION from the test properties.

Change-Id: I472dcbfa6897370aecddfa27302ea4c337f818a9