Replace newlines in sucessful test output.
In the CMake file for each test, the new lines of the output are removed to make
parsing easier.  Change the CommonFunctions.cmake to put the new line characters
back into the output if the test passes the XINDEX function.  This makes the
output on the dashboard easier to read.

Change-Id: I43f27ddd6bdcbc2504d6c687f9ea1584c598ea0a