BUG: Fix  MakeCommand not passing correctly
The script runs of the testing were not correctly capturing the MAKECOMMAND
set in the CMakeLists.txt. The creation of the CTestCustom.cmake.in passes
the configured variable into the script runs by setting the CTEST_BUILD_COMMAND
to the MAKECOMMAND that is set by the configuration. This step is done as the second
to last step of configure, followed only by including the CTest package. Remove the
return from the Cache part so that these final steps are reached. Remove the explicit
setting of the command from the vista_common.cmake

Change-Id: I376f6006753b619ce7beadcdf957f35a33888d98