Change Update step to report both repositories
The update step of the vista_common.cmake only reported the Testing Suite
changes on the dashboard.  Change the  update steps include an
execute_process to check out  and update the VistA CodeBase master branch
repository and append the result file to the notes which are submitted
with the build in addition to the regular update of the testing repository
Also, put the SHA key of the repository from after the update has
been made in the note.Only do this step if the files have been created
already.

Change-Id: Ic4f42f385b7d88480db5e117bee8de974103ef47