Update error reporting to include line numbers
Change the ReportXINDEXResult function in the CommonFunctions.cmake file to
find the tag of the function that contains the error.

Use the name to find the starting position via a grep in the routine file,
then use the position to find the line that contains the error

This lookup should only occur if the grep executable has been set and the
error has a tag entry.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-25
Change-Id: I1b9e111a65c5959508ef52b16ae17ec9b9c4cb0d