Update find_program to find_Package
The find_program commands that are used currently only give *.exe as the name of
the program to find.  The find_package command is a more robust command used to
find and capture external packages. The find_programs for Git and Python in each
of their section should be replaced with a single find_package for each program.
This can be done at the very top of the CMakelists.txt file.

Doing this requires a change in the variable names used in the rest of the
testing suite. The Python variable changes from PYTHON_EXEC to
PYTHON_EXECUTABLE.  The same change is done for the Git variable.

Suggested-by: Christopher Edwards <christopher.edwards@krminc.com>
Change-Id: I25fd721adc83c0d345c6a572bddca96a1e3b23c6