Move XINDEX Testing Framework to Python
This commit changes the environment of the XINDEX test drivers from
expect/Cache scripts to Python.  This change will allow greater
flexibility in the testing, as the python scripts can be used to a greater
degree than either of the previous scripts.

The method of communication to the instance is still the same,
Windows Cache is accessed via Telnet and Linux Cache and GT.M are accessed
with an expect program, though it is written in Python.  Pexpect,
http://sourceforge.net/projects/pexpect/files/ ,is an external module to
python, and needs to be installed prior to use.  But once installed, it
can be imported like any other library in Python.

To install Pexpect, unzip the download and run: 'python setup.py INSTALL'

The PythonRoutine.cmake.in is used to call the python scripts that are
created at configuration time.  The running of the tests remains the same.

Change-Id: Ib868e18fefe8f383e9e2ede0ce1bb1f185f8a91a