Include pexpect in OSEHRA Testing Repository
Add the python files that are contained in the
pexpect tar file. They are included to eliminate the need to
install the package before use.  Change the  CMakeLists.txt file
such that in the scripts where pexpect is used,a command is included which
adds this Python directory to the system path.

Change-Id: I1c81a70ad21a60dc20c8871a1690ce77ccb2ed48