Fix spawn command for Cache on Linux
The script command for the Cache on Linux needs to utilize
a "-U" argument to specify what namespace to start the instance in.
The python scripts are written to expect the user specified namespace as
the first prompt after signing in.  The current spawn command will present
the script with the USER namespace.

Change-Id: Ic29a9113440f118a587735bacd5c9c0822c5f776