Add CMakeLists.txt files to configure Delphi 2007 projects
Generate the project files instead of hard-coding them in the source
tree.  This enables multiple out-of-source build trees that do not touch
the original source at all.  Also add "make" rules to drive the project
file builds using MSBuild.

Currently the generated projects work automatically for command-line
builds with Delphi 2007.  The projects can be loaded in the IDE but the
user will have to manually set the PATH to point at the BDK output Bpl
directory and load the Delphi components.