Update Nightly import to run from python
Change the CMakeLists.txt to have the configuration of the Nightly import
scripts to have both systems utilize the same configuration. Move this
below the XINDEX test creation to utilize the same variables that are set
in the system identification blocks. Merge the two options of "clean_cache" and
"clean_database" into a single "clean_database" option that configures the
Python and CMake files when needed. Change the pexpect spawns to include
the 'timeout=None' argument, which eliminates the timeout while waiting
for the import to finish. Create the option for the user to choose to
supply a username and password for an instance that requires a login.

Change the import scripts (Routine,global and Initialize) to utilize the
OSEHRAHelper module as the method of connection to the MUMPS instance.

Change the OSEHRAHelper module to use an non-optional namespace argument
and have a check in the initialization to see if a string is provided.
If none is provided, use a default name.

Change the ImportRG.cmake to have 'if' statements that separate the GT.M
removal steps of the old environment from the Cache steps. It also calls
the configured python files that are created in the configure step.

Change the 'if(CControl AND CTerm)' statment to only depend on the CControl
variable.  CTerm is not available on Linux/Cache. Append Linux paths
to the "_Cache_Path" variable so that the Linux environment paths are included
in the search for the ccontrol program.

Changes to the CMakeLists.txt file for testing with Cache on Linux
were adapted from Christopher Edwards (Change I33e749dc).

Change-Id: If880e7e0dfd723fd5a676b10e9b8aed220b1e69b