Create and Utilize OSEHRAHelper Python module
Create a python module called OSEHRAHelper that is initialized with data
about the instance you will connect to. The Module  will automatically
determine the type and connection method.

Based on the system it will use the telnetlib to connect to a Win32-Cache
instance, and use pexpect to connect to the two Linux types.  With this module,
the scripts for testing are more generic, utilizing an abstract script to
connect to all three MUMPS instances. It will also store information which
can be used for code that needs a particular platform to run.

The information needed during initialization is the path to
a log file(required) and optionally: an Instance name, a namespace and
a location.  The optional arguments have defaults which will be used if
no variable is provided.  'Instance' defaults to 'CACHE', namespace to
'VISTA', and location to '127.0.0.1'.

Create two new options in the CMakelists.txt file to customize the XINDEX
scripts.  One option for needing to log into the instance called
"USE_LOGIN", the second for changing namespace at the beginning of the
script.  The USE_LOGIN will ask the user for the username and password to
supply to the instance when necessary.

Create a generic XINDEX python script that the testing will be called from the
PythonRoutine.cmake.in after configuration that takes 7 arguments.  The
arguments need to be in the following order: LogFilePath,
ListofRoutines (automatically generated), INSTANCE, NAMESPACE, USERNAME,
PASSWORD, VISTA_SOURCE_DIR.  The XINDEX python script should use both the
system type and the VistA.type to insert prompts that are specific to
particular instances.

Change-Id: I2de1dd75a7f0d353d77512c44d36ed9eddb4f015