Restructure OSEHRA Helper and XINDEXScript
Add a 'wait(PROMPT)' command to the XINDEXScript file just before the 'D
^XINDEX' command is written.  This eliminates the possiblity that the
command will be written before the system is ready to recieve it.

Add a line in the initialization of each connection type to have a
self.namespace be set. This allows the wait(PROMPT) to function correctly

Change-Id: I65692219317369da4aa4c90edb26b28dfe59a3bb