Automate functional testing of the CPRS GUI
This commit contains the script necessary to set up a base of information
needed in the functional testing of CPRS.  For each system, a
Clinic,Doctor,Nurse,Patient are created and the environment is set for
CPRS 28 and Vitals manager to be used with the installation.  It also
starts the TCP listener on port 9210 for communications in the cache
instances. The scripts are designed to be used after importing the routines
and globals from the OSEHRA repository

The FunctionalTesting.sikuli folder contains the necessary screenshots and
files needed to utilize the automation of GUI testing via the Sikuli
program

The cross-platform open-source program uses OpenCV to find and act upon
objects in the screen.  It is necessary to install Sikuli prior to testing
and can be downloaded for free from www.sikuli.org.

The program does expect to act upon icons on the Windows Desktop. A zip fi
containing the programs needed can be found on the OSEHRA site at:
http://www.osehra.org/document/guis-used-automatic-functional-testing.

The CPRSPostImportScript will be configured on all types of instances, and
will create a test to run the script.  The Sikuli test is only added on
Windows platforms, due to CPRS only running in a Windows environment

Change-Id: Iedd2c9ff2c6636ae43d5af0cdea36caba02b1830