Expand MUnit Testing capabilities
Change the MUnit testing scripts to not use the platform dependent
programs, but to use the OSEHRAHelper module and Python to install the
KIDS patches and Import and Run the routines.

Change the CMakeLists.txt file to have a UnitTesting section which is
outside of any platform specific code which installs the KIDS patch on
configuration.  The section should also glob the UnitTest routines to
create the necessary import and testing scripts.

Create the generic python scripts which can be configured and renamed with
the correct package for testing. There should be one python file for
running the tests, a cmake file to drive the previous one, and a file for
the import of the unit test routines.

Change the OSEHRAHelper.py module to have a 'multiwait' function which takes
a list of strings to match. This new list is used in the expect command in
either system and returns the index of the expression that was matched.
If the object passed to the multiwait function is not a list,
the program raises an error.

Change-Id: Ie06809e5b84da5b3a2c662a701634da7828bec94