Allow user to choose failure conditions
Change the CMakeLists.txt file to have an option called
USE_XINDEX_WARNINGS_AS_FAILURES
that asks the user if they want to use XINDEX warnings as a failure
condition during testing.

Change the PythonRoutine.cmake.in to pass the configured result of the
USE_XINDEX_WARNINGS_AS_FAILURES option to the ReportXINDEXResults function.

Change the CommonFunctions.cmake to have an if statement to determine the
string that is used to determine a failure condition.  If the
USE_XINDEX_WARNINGS_AS_FAILURES is turned on, use the string "F -|W -".
If the option is off, just use "F -"

Change-Id: I603775575aa79b1e7aacc0c63f32985724f39959