Change Prompt handling in OSEHRAHelper
Change the OSEHRAHelper.py to have a new variable set in the intialization of
the connection to VistA, self.prompt.  For Cache, this will be set to the
 current namespace with a '>' character appended onto the end. On GT.M, the
program will try to use the value from the environment variable "gtm_prompt".
If this doesnt exist, the value "GTM>" is used.

Change the wait functions for all three of the connection types to return
self.prompt in the event that given command is "PROMPT".

Change-Id: I2293f1d2f4879bd2fe51bb26efe060a4f2dee1c9