Fix MUnit install logic within ImportRG.cmake.in
Change the IF statement which controls whether the MUnitInstallationScript
is run to no longer use length as the metric.  If the file is not found
automatically, it will still have a length even though the path isn't
there.

Instead, change the statment to check if the value given for
MUNIT_KIDS_FILE is a valid file and if the user has turned the
INSTALL_MUNIT option on.  Set a local variable equal to the state of
INSTALL_MUNIT to avoid passing the value "ON" as an argument to the if
statement.

Include a message to warn the user if both AUTOMATED_UNIT_TESTING and
CLEAN_DATBASE options are turned on, while the INSTALL_MUNIT option is
turned off.  This combination of options will cause all MUnit tests to
fail in the event of a dashboard submission, due to the rewriting of the
database without reinstalling the MUNIT KIDS file.

Change-Id: I00916e566ba3e4a26b286c902988fda0e198a34e