Add CHKCUR^DGPTCO1 Unit test
Add a UnitTest routine that checks the CHKCUR entry point of the DGPTCO1
routine which has an if statment which contains an Non ANSI operator.
If the line remains unchanged, the GT.M implementations will cause a MUMPS
error to be thrown during execution while the Cache will not.

Create a new directory underneath the UnitTest in the source directory which
contains log files for the UnitTests.  This file will be used as the
baseline output to check for changes in functionality.

Change the UnitTest.cmake.in to go through each .M file in the directory
and check for a baseline file with the same name.  If it exists: clean the
test output of command characters and semicolons, then try to remove the
baseline text from the test output.  If nothing is removed, then the
baseline text isnt found in the test output, and the test fails.

Change the CommonFunctions.cmake files to have additional lines checking
the output file for the message "encountered (1-9) error".  This line
checks for the M Errors and should cause the test to fail.

There are also MUnit tests in the test cases to make sure that the
IENs of the globals are maintaining consistency.  The GETDGIEN entry point
should always return DGIEN=22.

Issue Tracker:       http://code.osehra.org/issues/view.php?id=2
Code Change to Test: http://review.code.osehra.org/#/c/168/

Change-Id: I77e1033f3502f95fea69b7158636ce930720256e