Introduce DUnit testing to CPRS
Create a new directory on the top level of the VistA-FOIA called
CPRSTesting which contains a sub-project with the same name.

Create a CMakeLists.txt in the 'Testing/CPRS' directory which adds the
CPRSTesting via the add_delphi_project macro.  It should also add tests
based on the file names in the Tests directory.

Create a pascal file called UnitTest.pas with two functions, one
(addSuite) to register a suite of tests passed to it as an arguement
while storing the name and the suite in two arrays. The second, called
runTests, will take an argument string and look to match the name of a test
suite in the name array.  If one matches, run that suite of tests, else it
should write a message saying "No registered tests match the supplied
name".

The CPRSTesting project is very straightforward.  If an argument is passed
to it at runtime, it calls runTests with the argument being passed on
further.  If no argument is passed, all registered tests should be run.

The test files are kept in 'Testing/CPRS/Tests/'.  These file names will be
used to create the string passed to the executable.  The name of the file,
without the extension, will append 'Tests' and that will be the suite that
is searched for.  For example,  UTSignonCnf.pas contains a suite named
UTSignonCnfTests.

OSEHRA-Issue: http://issues.osehra.org/browse/VAF-60
Original-Change-Id: Ie68f7caff3a9b74cc2c3e234ac40402f052e838b