Allow CPRS F.T. to call executable directly
Take the FunctionalTesting.py and make it a configurable file that will
start the CPRSChart/VitalsManager executable without clicking on an icon
on the desktop.

Also, take the FunctionalTesting.html and make it a configurable file to
match the FunctionalTesting.py

The script should check to see if an icon exists for the program, if it
doesn't find the icon, it should try to start the program using the
openApp() function of Sikuli.

Add a corresponding line to the files which adds a ImageSearchPath which
points back to the original FunctionalTesting.sikuli folder using the
CMake variable 'sikuli'.

Change the Win32 specific part of the FuctionalTesting in CMakeLists.txt
to now search for CPRSChart.exe and VitalsManager.exe, along with having
cached variables for the port that the Listener has been started on and
the server address of the instance.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-10
Change-Id: Ie6cbb3e3488b2965f38389bd78f57bb43bd5508d