Change the method of creating the Clinic
Change the PostImportSetupScript to add the VistA Health Care clinic
through the SDBUILD menu option instead of FileMan.  This is due to the
necessity of setting up availability hours for appointments in the clinic.
The clinic should now have 1/2 hour appointements slots from 8am-12pm, and
from 12:30pm-3pm

This should be added to the end of the section where the System Manager is
in the ZU menu.

The information on the SDBUILD menu was found here:
http://medsphere.org/thread/1382

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-33
Change-Id: I89cd9753bc442e2268ca0429c79f7dc3b43ee9f8