Change CPRS setup to ImportRG
Move the CPRSPostImportSetupScript.py.in out of the FunctionalTestingSuite
folder and remove the 'CPRS' from the name.

Remove the CPRSPostImport Test from the CMakeLists.txt and move the
configuration of the newly named PostImportSetupScript to fall under a new
option called VISTA_SITE_SETUP, that asks the user if they want to
populate the instance with fictional information.

Add a copy of ZTLOAD1.m to a directory called 'Patches'.  This file should
follow the ZTLOAD1 that is found in the patch in this entry:

http://review.code.osehra.org/#/c/187/

This file should be copied in place of the current ZTLOAD1 in the OSEHRA
Directory during the run of ImportRG but prior to the packing of the
routines.ro file on GT.M only.
This will ensure that the PostImportSetupScript will complete
successfully on GT.M.

Change the SystemTesting if conditions from (VISTA_CPRS_FUNCTIONAL_TESTING)
to (VISTA_CPRS_FUNCTIONAL_TESTING AND VISA_SITE_SETUP).  The CPRS testing
requires the sample data.

OSEHRA Issue: http://issues.osehra.org/browse/OAT-16

Change-Id: Iee12a4944729181253766fb460ca273eeb372099