EHN: Modified OSEHRAHelper.py to support timeout
Also added support to be able to store the text output after the wait like pexpect module

Modified GlobalImport and RoutineImport to use the new interface.

Also make sure that log is written to disk in case of the interrupt

Depending on the IO speed of the machine, the process of importing
globals can timeout in some cases. Here we make this a configurable value in CMake.

Change-Id: I5b3582f39906de161bfbd61a88c89401576a375b