Update Problem List RAS Test
The 2012-08-24 FOIA introduced MSV-IV codes into the Problem List.  One of
the codes used also has a MSV-IV code at that number.

Change the catad function in PLActions.py to contain a multiwait with the
'Ok?' and 'Select 1-5' strings.  If the 'Select 1-5' string is found,
select the first option and have 4 returns in the write string.  Otherwise, the
original write statement should be used.

Change-Id: I9371427195b2908f2133d3ce82f43b417e6db8c0