Further updates to RAS_Pl_test
Remove the '?' from the multiwait in the catad function.  It is a regular
expression special character.  It causes the first string to match either
'O' or 'Ok', which always matched due to the capitalized 'STOP' in the
output.

Change the second potential match to "STOP or Select' to remove the number
range dependency.

Add another return to the end of the write command within the if statment.

Change-Id: Ib04671e9fd1bae51ae8d729fcb77ee11c1935a8c