Update CPRS Functional Testing suite
Update the screenshot for the "VISTA HEALTH CARE" selection when creating
a visit in CPRS.  With the new clinics, it is no longer the first option
in the list.

Retake the screenshot to remove all traces of the border in the upper-left
corner.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-26
Change-Id: Ia62e4149ba6235a5f29c0f4c6c9a6049f7887886