Update PostImportsetupScript for new FOIA
With the release of the 26-08-2012 FOIA, the PostImportSetupScript failed
while running on GT.M. Cause was determined to be that the accompanying
Volume Set changed from FOI to PLA, the value we used.

Change the PostImportSetupScript to no longer use a long string to set the
volume set, but to have a multiwait.  The multiwait should determine if
the volume set is new ('Are you adding') or old ('VOLUME SET') and answer
accordingly.

JIRA Issue:  http://issues.osehra.org/browse/VAF-18

Change-Id: If43e3e9a024d86020ea5f0562fb9f4677f1ef83a