Update Patches directory
Change the file in the patches from a .m file to a .ro file.  This removes
the need to copy the file into the source directory prior to collection of
routines.

Add an 'if' statment after the routine import step of the ImportRG.cmake
to import the ZTLOAD.ro file, as it needs to be done before the ZTMGRSET.

Add the configure_file step to the CLEAN_DATABASE block to configure
routineimport.py.in to import the ZTLOAD.ro file.  Set the RoutineImportFile
to the path to the ZTLOAD.ro file before configuring.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-23
Change-Id: If90903eeb3807acd4bc2d7583d3c767d194a5d40