Write instructions to download VA VistA FOIA patches
Add file 'Scripts/HowtoDownloadPatches.rst' containing instructions to
download patch files from the VA VistA FOIA web site.  Store the files
in the 'FOIA/' directory mirroring the web site layout.

Script the download in order to select files with high granularity and
to make it reproducible.  Be sure not to hammer the server.