Introduce coverage using CTest
As of CMake 2.8.9, CTest has the capability to process MUMPS coverage.
Using one of two platform specfic functions( %SYS.MONLBL on Cache and
VIEW:TRACE on GT.m), CTest can parse the output to show coverage on the
Dashboard.

Add an option (USE_COVERAGE) in the CMakeLists.txt to ask the user if they
want to run coverage.

Add if statments to check for USE_COVERAGE to write out the coverage path
file.

  packages: ${OSEHRA_PATH}/Packages
  coverage: ${VISTA_BINARY_DIR}/Testing/Temporary/Coverage
The GT.M should have the file extension '.mcov' and the Cache file should
have a '.cmcov' extension.

Add 'startCoverage' and 'stopCoverage' functions for each of the sections
in the OSEHRAHelper.

The GT.M version should start the trace and save it
into a global called '^ZZCoverage'.  The stop function should stop the
trace and write the global out in the ZWR format.

The two Cache versions should start the monitor for all routines and
should monitor the lines (option 2) on all processes (option 1).  The stop
function should write out the coverage of all routines in a CSV file
(option 6).

All of the 'stop' functions should write out the file into the
"${VISTA_BINARY_DIR}/Testing/Temporary/Coverage/" directory.  Each should
have the file extension that matches the one created above.

Add the steps to the UnitTest.py.in and the XINDEXScript.py to call these
functions if the USE_COVERAGE value is 'ON'

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-17
Change-Id: I65fc0e54c75f993d0f7b3a8f5937c69f63e659bb