Add init script for Caché
Caché runs as a daemon that needs to be started and shutdown cleanly
with the operating system.

This script provides start, stop, and status for a Caché instance named
CACHE. It also assumes that there are no other Caché instances running
on the system as the lock file is statically named.