Update PostImport to alter default HFS path
Add code to the PostImportSetupScript to go into Fileman and edit the
DEFAULT HFS DIRECTORY entry in the KERNEL SYSTEM PARAMETERS file.  This
entry causes problems when temporary files are created by the system.

Add a CMake variable for the user to set the path to a vaild temp
directory.  If no change has been made, use the '@' to delete the entry in
the instance, which will use the working directory of the process as the
temporary directory.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-36
Change-Id: Iebae40b7c94942cbf431df9704a6d66ed7072265