Update to the UnitTest generation
Add a TestList file to each of the packages within the
UnitTest/VistA-FOIA/Packages/ directory. The name (without extension) of each
routine that contains MUnit tests should be added to the when adding a new
routine to test.

This is to eliminate the error that could be thrown when attempting to run
a routine that doesn't have any MUnit tests. It also allows for setup
routines to be placed alongside the tests and not have them be tested.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-45
Change-Id: I79c0007bf51634dd06c09b2203419212a7dcb2fb