Update Repository Directory usage
Change the paths for logging and coverage to use the TEST_RESULTS_DIR as
the basis for where log files and coverage files will be placed.

Combine the two coverage options in a separate section at the bottom of
the CMakeLists.txt intead of being in the platform-specific code.

Put the combined code under an 'if' statment conditional upon having a
CMake version that is 2.8.9 or greater.

Move the TEST_RESULTS_DIR variable from FunctionalTest/CMakeLists.txt to
the main repository CMakeLists.txt and make the default value of the
variable the path to the "VISTA_BINARY_DIR/Testing/Temporary" directory.

Change the paths given to the XINDEXScript and the MUnitScript path from
the "VISTA_BINARY_DIR/Testing/Temporary" to the TEST_RESULTS_DIR

Change the path given to the .[c]mcov file that determines the directory
that the coverage files are stored in from
"VISTA_BINARY_DIR/Testing/Temporary/Coverage" to
"TEST_RESULTS_DIR/Coverage"

OSEHRA-Issue: http://issues.osehra.org/browse/REF-95
Change-Id: I91c9dfc5a91b7fd87e83a2480fa273b09b606b64