Organize OSEHRA-Automated-Testing Directory Structure
Move the folders and files around within the OAT to prepare for a merge to
a new, more general repository.

Removed unused files :
Nightly.sh.in, GTMEnvironment.sh.in, FunctionalTest/CPRSTesting.cmake.in

Move the CommonFunctions.cmake to a top level CMake/ directory.

Rename the libs directory to Python, while maintaining the internal
structure.

Move the top level CMakeLists.txt file to the Testing directory
and create a new top level CMakeLists.txt which simply adds
the Testing directory and the Utilities/Dox subdirectories based upon
options.  Change the paths in the Testing/CMakeLists.txt to no longer
refer to the VISTA_SOURCE_DIR but instead to a CMAKE_CURRENT_SOURCE_DIR,
except for when referring to the other top level directories.

Move the UnitTest and RAS tests that are specific to a package to a directory
under a Packages/ directory called Testing. This Testing/ folder can
contain an MUnit director and/or a RAS directory to hold the test files.

Move the XindexExceptions files to a XINDEXException folder within their
respective folders under the Packages/ directory.  Each file
$package.$routine file should be renamed to $platform.routine:
For example:
  Cache/Kernel.XGKB -> Packages/Kernel/XINDEXException/Cache.XGKB.

Create a Setup directory under the Testing folder, which holds the
ImportRG.cmake file, the ZTLOAD.ro file, and all accompanying python
files.

Move the FunctionalTest folder to be named Functional and have it under
the Testing directory.  This will contain the CMake and python files for
both RAS and Sikuli testing, but only contain the source files for the
Sikuli testing.

Create a Utilities directory at the top level. Move the Dox folder into
this directory and add a CMakeLists.txt to the directory.  Move the
"ROUTINE_CALL_GRAPH" option from the Testing/CMakeLists.txt to the top
level.  If the option is chosen, add this "Utilities/Dox" folder as a
subdirectory.

Move the XINDEXExceptGen.py file from the Scripts directory to an XINDEX
folder within the Utilities directory.

Update all the testing scripts to use this new layout.

Change-Id: Ic7f251c42d8012058ae9449296ee71d85da2ba5b