Factor out journaling capabilities from create Vista Instance
Journaling can be enabled or disabled depending on the need. During
initial import of VistA it should be disabled as you will use 2x disk
space for the journal files and database.

Also, currently the OSEHRA Test Harness will disable journaling when it
resets the database (even if you had it enabled to begin with) as the
setting is stored in the .dat not in the .gld (the gld is the only thing
kept by the test harness). By separating the enabling and disabling to
two scripts journaling can be enabled/disabled at will.