SC, PL, and REG function test updates
Add updates to the current Problem List and Scheduling RAS tests.

Add a new set of RAS tests for the Registration package that tests the ADT
functionality.

Add a new actions python file for interacting with the CPRS menu.

Add new patient data to the patdata0.csv

Rename the importUsers.py.in to importPats.py.in.

Change-Id: Ie4c059376a270a1a96500d46909e18385b0d1ce4