Rename OSEHRA VistA-Patches to OSEHRA VistA
The history behind this commit was originally developed in a separate
VistA-Patches repository.  It was filtered to remove the actual VA FOIA
VistA patch files but keep the OSEHRA scripts and documentation intact.

Prepare to publish this history in a new OSEHRA "VistA" repository that
will also integrate other OSEHRA VistA tools and infrastructure.  Update
the repository name in README.rst and the remote repository names in
Scripts/GitSetup/config.  Update the NOTICE file project name and
copyright years.