CMake: Allow Delphi and Testing components to build together
The CMake infrastructure merged in from OSERHA-Automated-Testing hacked
the values of CMAKE_MAKE_PROGRAM and MAKECOMMAND in order to allow
testing to run without a real make tool.  Generalize this approach and
use it only when a real make tool is not available.

When a real make tool is available, add custom targets to run the CMake
scripts like ImportRG during ctest_build.  Otherwise, collect a list of
such scripts and drive them under ctest_build using a MAKECOMMAND that
runs "cmake -P".

While at it, move the code to hide CMAKE_INSTALL_PREFIX up to the top
since we currently have no install rules.