Check for TEST_VISTA_FRESH variables during configure
There is currently no check that the variables that appear after
selecting the TEST_VISTA_FRESH option have been defined prior to use.

Add a check for the variables by adding the variable to a list,
looping through the list to visit each, and check that it is defined.
If it is not, display an error to say that the variable needs to be defined
before configuring.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-53
Change-Id: I865959adb805dc566a4ce492cbe919114fd12c5b