Remove XINDEX testing dependency on VistA-FOIA repo
Remove the dependency of the VistA repository on the VistA-FOIA repository
for creating the XINDEX scripts used in testing.

Move the set of the TEST_VISTA_FRESH_M_DIR to the UnitTest setup section,
as those tests still need the Scripts Directory.

Use the ParseCSVforPackagePrefixes code to parse the Packages.csv file for
the prefixes and pass to the XINDEX test.

Change the places where the 'Routines' directory is globbed for routines to
test to instead use the ^%RD to query the instance for the routines using
the prefixes above.  This list of routines should be written to a
file and read in during the test to create the listing.

Change the XINDEXScript.py to be a configurable file.

Change the ReportXINDEXErrors macro in CommonFunctions.cmake to no longer
require the VistA-FOIA path as an arugment.  Change references to
${OSEHRA_PATH} to refer to ${ARGV6}, thus making it an optional parameter.

Change the if(GREP_EXECUTABLE AND TAG) statment to also check for the
ARGV6 value.  If it is defined, then proceed with finding the line number
of the error.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-55
Change-Id: I09ac8f852fd4c68a605a68e8e26c7cef240fe67b