Match XUP menu prompts after XU*8*593
The XU*8*593 patch has added a "<TEST ACCOUNT>" string to the XUP/ZU
menu option names. Change prompts within the InstallMunitKIDSBuild.py.in
and the PostImportSetupScript.py.in to prepare for the menu name change.

Remove all "select" or "option" strings from the wait strings, leaving
only the name of the menu that you are going to access.  This will
reduce the chance the "<TEST ACCOUNT>" will interrupt the search
string.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-46
Change-Id: Ib1ea58c5a4dbf5a11b0847b20da5c17f0c35e738