Change RAS Action script Menu strings
The XU*8*593 patch has added a new <TEST ACCOUNT> tag to the option in
the ZU and XUP menus.  Update the strings that wait for a menu name in the
RAS tests to not use "Menu Option" but instead wait for the name of the
menu.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-46
Change-Id: I0079d9cf0033ffd8cdb13824c52e20f10eb37c0e