VistATestClient: Enhance to better handle CTRL+C
In situation that CTRL+C is pressed, code to close VistA connection
might gracefully might be blocked and leave a zombie external process.
Modify to terminate external process right away.

Also modify to catch all exceptions instead of just TIMEOUT exception so
that we still have a chance to terminate the external process.

OSEHRA-Id: http://issues.osehra.org/browse/VAF-81
Change-Id: Iba7a9c34b3d74165059528e360af8720345aa845