Python: Add PackagePrefix parsing library code
Add a file ParseCSVforPackagePrefixes.py which contains the code to read
a csv file like the Packages.csv and parse the Package and prefix
information and return a sorted list of prefixes.

Add a test data .csv file and a test which uses that data and compares it
to a hardcoded value.

Add a README.rst file to explain the code and give examples of output.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-55
Change-Id: Ic784b6983ab68e40bed1ba4d15078d1f2a3ee355