Problem List: Add test and revert patch check
Add new file ZZRGUTEX to the TestList file for inclusion in the MUnit test

Restore the 'Q:$T(^GMPLAPI2)=""' line back to the beginning of the file in
the ZZRGUT4. This line causes the test not to run unless the refactoring
patch is installed.

OSEHRA-Issue: http://issues.osehra.org/browse/OAT-60
Change-Id: I3df501fd858d70416d30a29882ad6a574d2a8c1f