KIDSPatchSequenceApply: make __reloadPackagePatchHisty more lenient
In the process of installing CPRS 29 KIDS patch CSV_HDS_1_0.KID,
the _reloadPackagePatchHostory_ method throws an assertion error as
"CSV HDS" is not a valid namespace. The original code assumes the any
installation name in the format like "ABCD 2.0" implicitly indicates
that ABCD is the namespace and 2.0 is the version #, this in general is the case.
However, in the case, the code needs to be more lenient to handle it gracefully.

KIDSPatchOrderGenerator: add more debugging message for multibuilds dependencies.

OSEHRA-Id: http://issues.osehra.org/browse/VAF-86
Change-Id: If9c03bf025cdadaa329144cf940b840e1acef957