Update Doc and Import for removal of VistA-FOIA/Scripts
Update the import documentation and the ImportRG utility to prepare for
the deletion of the VistA-FOIA Scripts directory.

Add a second call to PackRO.py to capture ZGO/ZGI from the Scripts
directory of this repository

Add the proper configure_file commands to the CMakeLists.txt file in the
Testing directory.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-64
Change-Id: Ic1f80e2ecaddc84df0dce1e837751aad86964a84