PatchOrderGenerator: Update patch seq # dependency prior to multibuilds
Update patch dependencies based on sequence number before updating
patch dependencies for multi-builds, so that multi-builds patch won't
be potentially out of sequence order.

Also clean up logic that is obsolete, enhance the algorithm to identify
starting node set for topologic sort, set csv dependency of the first
build of multi-builds to be the same as the second build if first build
does not have one.

Change-Id: I2f3ed0831275bea2bc0093153823693413247fb5