RAS: Update wait strings for reg_test004 output
The reg_004 test in the Registration RAS waits for hard coded numeric
values in the printout of information about stent replacement.

Update the wait strings that are only numbers to use a regular expression
that matches both the information string and the number.  For example:

'392'  ->  'Diagnosis Related Group: +[0-9]+'

Update the wwgeneric function in the ADTActions.py library
to use the regular expression wait ('wait_re') instead of the normal
wait function.

OSEHRA-ID: http://issues.osehra.org/browse/OAT-67
Change-Id: Icc338b0beffdebb0bdd7888e0d9373694f792065