Sikuli: Update Sikuli click/type commands and visit screenshot
Change all type and click commands to use a new local function that has
'waits' for the passed image before performing the action.

Add new screenshots for creating a visit by the user.  The previous
image used the doctors name for both cases, when CPRS prints out the
name of the person who just signed in.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-68
Change-Id: I74b68c1f8386db25becb1da6ce2077d5d67e7e2c